+++ benchee/regex.exs
@@ -1,4 +1,5 @@
 Benchee.run(%{
   "Unicode"  => fn -> Unicode.Property.uppercase?("A") end,
-  "regex" => fn -> Regex.match?(~r/\p{Lu}/u, "A") end
+  "unicode regex" => fn -> Regex.match?(~r/\p{Lu}/u, "A") end,
+  "regex" => fn -> Regex.match?(~r/[A-Z]/u, "A") end
   })